+++ src/aurora/gff3file.cpp
@@ -140,7 +140,7 @@ void GFF3File::loadLists() {
 
 	// Counting the actual amount of lists
 	uint32 listCount = 0;
-	for (uint32 i = 0; i < rawLists.size(); i++) {
+	for (size_t i = 0; i < rawLists.size(); i++) {
 		uint32 n = rawLists[i];
 
 		if ((i + n) > rawLists.size())
@@ -155,7 +155,7 @@ void GFF3File::loadLists() {
 
 	// Converting the raw list array into real, useable lists
 	uint32 listIndex = 0;
-	for (uint32 i = 0; i < rawLists.size(); listIndex++) {
+	for (size_t i = 0; i < rawLists.size(); listIndex++) {
 		_listOffsetToIndex[i] = listIndex;
 
 		const uint32 n = rawLists[i++];
@@ -297,7 +297,7 @@ Common::SeekableReadStream &GFF3Struct::getData(const Field &field) const {
 
 // --- Field properties ---
 
-uint GFF3Struct::getFieldCount() const {
+size_t GFF3Struct::getFieldCount() const {
 	return _fields.size();
 }
